Removed +1 after reading an element into memory
The +1 seems to be for handling the capacity of the vector which may increase as well when elements are added to the vector, to keep some "breathing room" for future elements. However, this behaviour is compiler dependent so there is no point in adding +1 explicitly as it may not be correct.